+++ scripts/add_electricity.py
@@ -964,7 +964,11 @@ def attach_hydro(
 
 
 def attach_existing_batteries(
-    n: pypsa.Network, costs: pd.DataFrame, ppl: pd.DataFrame
+    n: pypsa.Network,
+    costs: pd.DataFrame,
+    ppl: pd.DataFrame,
+    extendable_carriers: dict,
+    max_hours: float,
 ) -> None:
     """
     Add existing battery storage units from powerplants.csv to the network.
@@ -977,6 +981,10 @@ def attach_existing_batteries(
         DataFrame containing technology costs.
     ppl : pd.DataFrame
         Power plant DataFrame.
+    extendable_carriers : dict
+        Dictionary of extendable carriers for different component types.
+    max_hours: float
+        Amount of time it takes to fully charge batteries from empty if done at maximum power rate.
 
     Returns
     -------
@@ -992,29 +1000,90 @@ def attach_existing_batteries(
     # Aggregate batteries by (bus, carrier, grouping_year)
     batteries_grouped = aggregate_ppl_by_bus_carrier_year(batteries)
 
-    max_hours = snakemake.params.electricity["max_hours"]["battery"]
+    if "battery" in extendable_carriers["Store"]:
+        battery_def = "Stores and Links"
 
-    n.madd(
-        "StorageUnit",
-        batteries_grouped.index,
-        bus=batteries_grouped["bus"],
-        carrier=batteries_grouped["carrier"],
-        p_nom=batteries_grouped["p_nom"],
-        p_nom_extendable=False,
-        p_nom_min=batteries_grouped["p_nom"],
-        p_nom_max=batteries_grouped["p_nom"],
-        capital_cost=costs.at["battery", "capital_cost"],
-        max_hours=max_hours,
-        efficiency_store=np.sqrt(costs.at["battery", "efficiency"]),
-        efficiency_dispatch=np.sqrt(costs.at["battery", "efficiency"]),
-        cyclic_state_of_charge=True,
-        marginal_cost=costs.at["battery", "marginal_cost"],
-        build_year=batteries_grouped["build_year"],
-        lifetime=batteries_grouped["lifetime"],
-    )
+        n.madd(
+            "Bus",
+            batteries_grouped["bus"] + " battery",
+            location=batteries_grouped["bus"],
+            carrier="battery",
+            x=n.buses.loc[list(batteries_grouped["bus"])].x.values,
+            y=n.buses.loc[list(batteries_grouped["bus"])].y.values,
+        )
+
+        # Add Stores for existing battery energy capacity
+        n.madd(
+            "Store",
+            batteries_grouped.index,
+            bus=batteries_grouped["bus"] + " battery",
+            e_cyclic=True,
+            e_nom_extendable=False,
+            e_nom=batteries_grouped["p_nom"] * max_hours,
+            carrier="battery",
+            capital_cost=costs.at["battery storage", "capital_cost"],
+            build_year=batteries_grouped["build_year"],
+            lifetime=batteries_grouped["lifetime"],
+        )
+
+        # Add charger Links for existing battery power capacity
+        n.madd(
+            "Link",
+            batteries_grouped.index.map(
+                lambda x: x.replace(" battery", " battery charger")
+            ),
+            bus0=batteries_grouped["bus"].values,
+            bus1=(batteries_grouped["bus"] + " battery").values,
+            p_nom=batteries_grouped["p_nom"].values,
+            p_nom_extendable=False,
+            carrier="battery charger",
+            efficiency=costs.at["battery inverter", "efficiency"] ** 0.5,
+            capital_cost=costs.at["battery inverter", "capital_cost"],
+            lifetime=costs.at["battery inverter", "lifetime"],
+            build_year=batteries_grouped["build_year"].values,
+        )
+
+        # Add discharger Links for existing battery power capacity
+        n.madd(
+            "Link",
+            batteries_grouped.index.map(
+                lambda x: x.replace(" battery", " battery discharger")
+            ),
+            bus0=(batteries_grouped["bus"] + " battery").values,
+            bus1=batteries_grouped["bus"].values,
+            p_nom=batteries_grouped["p_nom"].values,
+            p_nom_extendable=False,
+            carrier="battery discharger",
+            efficiency=costs.at["battery inverter", "efficiency"] ** 0.5,
+            marginal_cost=costs.at["battery", "marginal_cost"],
+            lifetime=costs.at["battery inverter", "lifetime"],
+            build_year=batteries_grouped["build_year"].values,
+        )
+
+    else:
+        battery_def = "StorageUnit"
+
+        n.madd(
+            "StorageUnit",
+            batteries_grouped.index,
+            bus=batteries_grouped["bus"],
+            carrier=batteries_grouped["carrier"],
+            p_nom=batteries_grouped["p_nom"],
+            p_nom_extendable=False,
+            p_nom_min=batteries_grouped["p_nom"],
+            p_nom_max=batteries_grouped["p_nom"],
+            capital_cost=costs.at["battery", "capital_cost"],
+            max_hours=max_hours,
+            efficiency_store=np.sqrt(costs.at["battery", "efficiency"]),
+            efficiency_dispatch=np.sqrt(costs.at["battery", "efficiency"]),
+            cyclic_state_of_charge=True,
+            marginal_cost=costs.at["battery", "marginal_cost"],
+            build_year=batteries_grouped["build_year"],
+            lifetime=batteries_grouped["lifetime"],
+        )
 
     logger.info(
-        f"Added {len(batteries_grouped)} existing batteries with total capacity "
+        f"Added {len(batteries_grouped)} existing batteries defined as {battery_def} with total capacity "
         f"{batteries_grouped.p_nom.sum()/1e3:.2f} GW (max_hours={max_hours})."
     )
 
@@ -1368,7 +1437,13 @@ if __name__ == "__main__":
     attach_hydro(
         n, costs, ppl, snakemake.params.renewable["hydro"]["hydro_min_inflow_pu"]
     )
-    attach_existing_batteries(n, costs, ppl)
+    attach_existing_batteries(
+        n,
+        costs,
+        ppl,
+        extendable_carriers,
+        snakemake.params.electricity["max_hours"]["battery"],
+    )
 
     if snakemake.params.electricity.get("estimate_renewable_capacities"):
         estimate_renewable_capacities_irena(